move file kartlegging and update relative imports
